+++ providers/aws/services/elbv2/check_extra717
@@ -1,0 +1,65 @@
+#!/usr/bin/env bash
+
+# Prowler - the handy cloud security tool (copyright 2018) by Toni de la Fuente
+#
+# Licensed under the Apache License, Version 2.0 (the "License"); you may not
+# use this file except in compliance with the License. You may obtain a copy
+# of the License at http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software distributed
+# under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+# CONDITIONS OF ANY KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations under the License.
+CHECK_ID_extra717="7.17"
+CHECK_TITLE_extra717="[extra717] Check if Elastic Load Balancers have logging enabled"
+CHECK_SCORED_extra717="NOT_SCORED"
+CHECK_CIS_LEVEL_extra717="EXTRA"
+CHECK_SEVERITY_extra717="Medium"
+CHECK_ASFF_RESOURCE_TYPE_extra717="AwsElbLoadBalancer"
+CHECK_ALTERNATE_check717="extra717"
+CHECK_SERVICENAME_extra717="elb"
+CHECK_RISK_extra717='If logs are not enabled monitoring of service use and threat analysis is not possible.'
+CHECK_REMEDIATION_extra717='Enable ELB logging; create la log lifecycle and define use cases.'
+CHECK_DOC_extra717='https://docs.aws.amazon.com/elasticloadbalancing/latest/application/load-balancer-access-logs.html'
+CHECK_CAF_EPIC_extra717='Logging and Monitoring'
+
+extra717(){
+  # "Check if Elastic Load Balancers have logging enabled "
+  for regx in $REGIONS; do
+    LIST_OF_ELBS=$($AWSCLI elb describe-load-balancers $PROFILE_OPT --region $regx --query 'LoadBalancerDescriptions[*].LoadBalancerName' --output text 2>&1 |xargs -n1)
+    if [[ $(echo "$LIST_OF_ELBS" | grep -E 'AccessDenied|UnauthorizedOperation') ]]; then
+        textInfo "$regx: Access Denied trying to list load balancers v1" "$regx"
+        continue
+    fi
+    LIST_OF_ELBSV2=$($AWSCLI elbv2 describe-load-balancers $PROFILE_OPT --region $regx --query 'LoadBalancers[*].LoadBalancerArn' --output text 2>&1 |xargs -n1)
+    if [[ $(echo "$LIST_OF_ELBSV2" | grep -E 'AccessDenied|UnauthorizedOperation') ]]; then
+        textInfo "$regx: Access Denied trying to list load balancers v2" "$regx"
+        continue
+    fi
+    if [[ $LIST_OF_ELBS || $LIST_OF_ELBSV2 ]]; then
+      if [[ $LIST_OF_ELBS ]]; then
+        for elb in $LIST_OF_ELBS; do
+          CHECK_ELBS_LOG_ENABLED=$($AWSCLI elb describe-load-balancer-attributes $PROFILE_OPT --region $regx --load-balancer-name $elb --query 'LoadBalancerAttributes.AccessLog.Enabled' |grep "^true")
+          if [[ $CHECK_ELBS_LOG_ENABLED ]]; then
+            textPass "$regx: $elb has access logs to S3 configured" "$regx" "$elb"
+          else
+            textFail "$regx: $elb has not configured access logs" "$regx" "$elb"
+          fi
+        done
+      fi
+      if [[ $LIST_OF_ELBSV2 ]]; then
+        for elbarn in $LIST_OF_ELBSV2; do
+          CHECK_ELBSV2_LOG_ENABLED=$($AWSCLI elbv2 describe-load-balancer-attributes $PROFILE_OPT --region $regx --load-balancer-arn $elbarn --query Attributes[*] --output text |grep "^access_logs.s3.enabled"|cut -f2|grep true)
+          ELBV2_NAME=$(echo $elbarn|cut -d\/ -f3)
+          if [[ $CHECK_ELBSV2_LOG_ENABLED ]]; then
+            textPass "$regx: $ELBV2_NAME has access logs to S3 configured" "$regx" "$elb"
+          else
+            textFail "$regx: $ELBV2_NAME has not configured access logs" "$regx" "$elb"
+          fi
+        done
+      fi
+    else
+      textInfo "$regx: No ELBs found" "$regx"
+    fi
+  done
+}
