+++ providers/aws/services/elbv2/check_extra79
@@ -1,0 +1,53 @@
+#!/usr/bin/env bash
+
+# Prowler - the handy cloud security tool (copyright 2018) by Toni de la Fuente
+#
+# Licensed under the Apache License, Version 2.0 (the "License"); you may not
+# use this file except in compliance with the License. You may obtain a copy
+# of the License at http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software distributed
+# under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+# CONDITIONS OF ANY KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations under the License.
+CHECK_ID_extra79="7.9"
+CHECK_TITLE_extra79="[extra79] Check for internet facing Elastic Load Balancers"
+CHECK_SCORED_extra79="NOT_SCORED"
+CHECK_CIS_LEVEL_extra79="EXTRA"
+CHECK_SEVERITY_extra79="Medium"
+CHECK_ASFF_RESOURCE_TYPE_extra79="AwsElbLoadBalancer"
+CHECK_ALTERNATE_extra709="extra79"
+CHECK_ALTERNATE_check79="extra79"
+CHECK_ALTERNATE_check709="extra79"
+CHECK_SERVICENAME_extra79="elb"
+CHECK_RISK_extra79='Publicly accessible load balancers could expose sensitive data to bad actors.'
+CHECK_REMEDIATION_extra79='Ensure the load balancer should be publicly accessible. If publicly exposed ensure a WAF ACL is implemented.'
+CHECK_DOC_extra79='https://docs.aws.amazon.com/waf/latest/developerguide/web-acl-associating-aws-resource.html'
+CHECK_CAF_EPIC_extra79='Data Protection'
+
+extra79(){
+  # "Check for internet facing Elastic Load Balancers "
+  for regx in $REGIONS; do
+    LIST_OF_PUBLIC_ELBS=$($AWSCLI elb describe-load-balancers $PROFILE_OPT --region $regx --query 'LoadBalancerDescriptions[?Scheme == `internet-facing`].[LoadBalancerName,DNSName]' --output text 2>&1)
+    if [[ $(echo "$LIST_OF_PUBLIC_ELBS" | grep AccessDenied) ]]; then
+      textInfo "$regx: Access Denied Trying to describe load balancers" "$regx"
+      continue
+    fi
+    LIST_OF_PUBLIC_ELBSV2=$($AWSCLI elbv2 describe-load-balancers $PROFILE_OPT --region $regx --query 'LoadBalancers[?Scheme == `internet-facing`].[LoadBalancerName,DNSName]' --output text 2>&1)
+    if [[ $(echo "$LIST_OF_PUBLIC_ELBSV2" | grep AccessDenied) ]]; then
+      textInfo "$regx: Access Denied Trying to describe load balancers" "$regx"
+      continue
+    fi
+    LIST_OF_ALL_ELBS=$( echo $LIST_OF_PUBLIC_ELBS; echo $LIST_OF_PUBLIC_ELBSV2)
+    LIST_OF_ALL_ELBS_PER_LINE=$( echo $LIST_OF_ALL_ELBS| xargs -n2 )
+    if [[ $LIST_OF_ALL_ELBS ]];then
+      while read -r elb;do
+        ELB_NAME=$(echo $elb | awk '{ print $1; }')
+        ELB_DNSNAME=$(echo $elb | awk '{ print $2; }')
+        textFail "$regx: ELB: $ELB_NAME at DNS: $ELB_DNSNAME is internet-facing!" "$regx" "$ELB_NAME"
+      done <<< "$LIST_OF_ALL_ELBS_PER_LINE"
+      else
+        textPass "$regx: no Internet Facing ELBs found" "$regx" "$ELB_NAME"
+    fi
+  done
+}
