+++ providers/aws/services/elbv2/elbv2_deletion_protection/elbv2_deletion_protection.metadata.json
