+++ providers/aws/services/elbv2/elbv2_deletion_protection/elbv2_deletion_protection_test.py
@@ -8,32 +8,32 @@ AWS_REGION = "eu-west-1"
 AWS_ACCOUNT_NUMBER = "123456789012"
 
 
-class Test_elbv2_listeners_underneath:
+class Test_elbv2_deletion_protection:
     @mock_elbv2
     def test_elb_no_balancers(self):
 
         from providers.aws.lib.audit_info.audit_info import current_audit_info
-        from providers.aws.services.elb.elb_service import ELBv2
+        from providers.aws.services.elbv2.elbv2_service import ELBv2
 
         current_audit_info.audited_partition = "aws"
 
         with mock.patch(
-            "providers.aws.services.elb.elbv2_listeners_underneath.elbv2_listeners_underneath.elbv2_client",
+            "providers.aws.services.elbv2.elbv2_deletion_protection.elbv2_deletion_protection.elbv2_client",
             new=ELBv2(current_audit_info),
         ):
             # Test Check
-            from providers.aws.services.elb.elbv2_listeners_underneath.elbv2_listeners_underneath import (
-                elbv2_listeners_underneath,
+            from providers.aws.services.elbv2.elbv2_deletion_protection.elbv2_deletion_protection import (
+                elbv2_deletion_protection,
             )
 
-            check = elbv2_listeners_underneath()
+            check = elbv2_deletion_protection()
             result = check.execute()
 
             assert len(result) == 0
 
     @mock_ec2
     @mock_elbv2
-    def test_elbv2_without_listeners(self):
+    def test_elbv2_without_deletion_protection(self):
         conn = client("elbv2", region_name=AWS_REGION)
         ec2 = resource("ec2", region_name=AWS_REGION)
 
@@ -56,26 +56,33 @@ class Test_elbv2_listeners_underneath:
             Type="application",
         )["LoadBalancers"][0]
 
+        conn.modify_load_balancer_attributes(
+            LoadBalancerArn=lb["LoadBalancerArn"],
+            Attributes=[
+                {"Key": "deletion_protection.enabled", "Value": "false"},
+            ],
+        )
+
         from providers.aws.lib.audit_info.audit_info import current_audit_info
-        from providers.aws.services.elb.elb_service import ELBv2
+        from providers.aws.services.elbv2.elbv2_service import ELBv2
 
         current_audit_info.audited_partition = "aws"
 
         with mock.patch(
-            "providers.aws.services.elb.elbv2_listeners_underneath.elbv2_listeners_underneath.elbv2_client",
+            "providers.aws.services.elbv2.elbv2_deletion_protection.elbv2_deletion_protection.elbv2_client",
             new=ELBv2(current_audit_info),
         ):
-            from providers.aws.services.elb.elbv2_listeners_underneath.elbv2_listeners_underneath import (
-                elbv2_listeners_underneath,
+            from providers.aws.services.elbv2.elbv2_deletion_protection.elbv2_deletion_protection import (
+                elbv2_deletion_protection,
             )
 
-            check = elbv2_listeners_underneath()
+            check = elbv2_deletion_protection()
             result = check.execute()
 
             assert len(result) == 1
             assert result[0].status == "FAIL"
             assert search(
-                "has no listeners underneath",
+                "has not deletion protection",
                 result[0].status_extended,
             )
             assert result[0].resource_id == "my-lb"
@@ -83,7 +90,7 @@ class Test_elbv2_listeners_underneath:
 
     @mock_ec2
     @mock_elbv2
-    def test_elbv2_with_listeners(self):
+    def test_elbv2_with_deletion_protection(self):
         conn = client("elbv2", region_name=AWS_REGION)
         ec2 = resource("ec2", region_name=AWS_REGION)
 
@@ -127,24 +134,34 @@ class Test_elbv2_listeners_underneath:
             DefaultActions=[{"Type": "forward", "TargetGroupArn": target_group_arn}],
         )
 
+        conn.modify_load_balancer_attributes(
+            LoadBalancerArn=lb["LoadBalancerArn"],
+            Attributes=[
+                {"Key": "deletion_protection.enabled", "Value": "true"},
+            ],
+        )
+
         from providers.aws.lib.audit_info.audit_info import current_audit_info
-        from providers.aws.services.elb.elb_service import ELBv2
+        from providers.aws.services.elbv2.elbv2_service import ELBv2
 
         current_audit_info.audited_partition = "aws"
 
         with mock.patch(
-            "providers.aws.services.elb.elbv2_listeners_underneath.elbv2_listeners_underneath.elbv2_client",
+            "providers.aws.services.elbv2.elbv2_deletion_protection.elbv2_deletion_protection.elbv2_client",
             new=ELBv2(current_audit_info),
         ):
-            from providers.aws.services.elb.elbv2_listeners_underneath.elbv2_listeners_underneath import (
-                elbv2_listeners_underneath,
+            from providers.aws.services.elbv2.elbv2_deletion_protection.elbv2_deletion_protection import (
+                elbv2_deletion_protection,
             )
 
-            check = elbv2_listeners_underneath()
+            check = elbv2_deletion_protection()
             result = check.execute()
 
             assert len(result) == 1
             assert result[0].status == "PASS"
-            assert search("has listeners underneath", result[0].status_extended)
+            assert search(
+                "has deletion protection enabled",
+                result[0].status_extended,
+            )
             assert result[0].resource_id == "my-lb"
             assert result[0].resource_arn == lb["LoadBalancerArn"]
