+++ providers/aws/services/elbv2/elbv2_desync_mitigation_mode/elbv2_desync_mitigation_mode_test.py
@@ -8,32 +8,32 @@ AWS_REGION = "eu-west-1"
 AWS_ACCOUNT_NUMBER = "123456789012"
 
 
-class Test_elbv2_request_smugling:
+class Test_elbv2_desync_mitigation_mode:
     @mock_elbv2
     def test_elb_no_balancers(self):
 
         from providers.aws.lib.audit_info.audit_info import current_audit_info
-        from providers.aws.services.elb.elb_service import ELBv2
+        from providers.aws.services.elbv2.elbv2_service import ELBv2
 
         current_audit_info.audited_partition = "aws"
 
         with mock.patch(
-            "providers.aws.services.elb.elbv2_request_smugling.elbv2_request_smugling.elbv2_client",
+            "providers.aws.services.elbv2.elbv2_desync_mitigation_mode.elbv2_desync_mitigation_mode.elbv2_client",
             new=ELBv2(current_audit_info),
         ):
             # Test Check
-            from providers.aws.services.elb.elbv2_request_smugling.elbv2_request_smugling import (
-                elbv2_request_smugling,
+            from providers.aws.services.elbv2.elbv2_desync_mitigation_mode.elbv2_desync_mitigation_mode import (
+                elbv2_desync_mitigation_mode,
             )
 
-            check = elbv2_request_smugling()
+            check = elbv2_desync_mitigation_mode()
             result = check.execute()
 
             assert len(result) == 0
 
     @mock_ec2
     @mock_elbv2
-    def test_elbv2_without_deletion_protection(self):
+    def test_elbv2_without_desync_mitigation_mode(self):
         conn = client("elbv2", region_name=AWS_REGION)
         ec2 = resource("ec2", region_name=AWS_REGION)
 
@@ -59,33 +59,30 @@ class Test_elbv2_request_smugling:
         conn.modify_load_balancer_attributes(
             LoadBalancerArn=lb["LoadBalancerArn"],
             Attributes=[
-                {
-                    "Key": "routing.http.drop_invalid_header_fields.enabled",
-                    "Value": "false",
-                },
+                {"Key": "routing.http.desync_mitigation_mode", "Value": "monitor"},
             ],
         )
 
         from providers.aws.lib.audit_info.audit_info import current_audit_info
-        from providers.aws.services.elb.elb_service import ELBv2
+        from providers.aws.services.elbv2.elbv2_service import ELBv2
 
         current_audit_info.audited_partition = "aws"
 
         with mock.patch(
-            "providers.aws.services.elb.elbv2_request_smugling.elbv2_request_smugling.elbv2_client",
+            "providers.aws.services.elbv2.elbv2_desync_mitigation_mode.elbv2_desync_mitigation_mode.elbv2_client",
             new=ELBv2(current_audit_info),
         ):
-            from providers.aws.services.elb.elbv2_request_smugling.elbv2_request_smugling import (
-                elbv2_request_smugling,
+            from providers.aws.services.elbv2.elbv2_desync_mitigation_mode.elbv2_desync_mitigation_mode import (
+                elbv2_desync_mitigation_mode,
             )
 
-            check = elbv2_request_smugling()
+            check = elbv2_desync_mitigation_mode()
             result = check.execute()
 
             assert len(result) == 1
             assert result[0].status == "FAIL"
             assert search(
-                "is not dropping invalid header fields",
+                "does not have desync mitigation mode set as defensive or strictest",
                 result[0].status_extended,
             )
             assert result[0].resource_id == "my-lb"
@@ -93,7 +90,7 @@ class Test_elbv2_request_smugling:
 
     @mock_ec2
     @mock_elbv2
-    def test_elbv2_with_deletion_protection(self):
+    def test_elbv2_with_desync_mitigation_mode(self):
         conn = client("elbv2", region_name=AWS_REGION)
         ec2 = resource("ec2", region_name=AWS_REGION)
 
@@ -140,33 +137,30 @@ class Test_elbv2_request_smugling:
         conn.modify_load_balancer_attributes(
             LoadBalancerArn=lb["LoadBalancerArn"],
             Attributes=[
-                {
-                    "Key": "routing.http.drop_invalid_header_fields.enabled",
-                    "Value": "true",
-                },
+                {"Key": "routing.http.desync_mitigation_mode", "Value": "defensive"},
             ],
         )
 
         from providers.aws.lib.audit_info.audit_info import current_audit_info
-        from providers.aws.services.elb.elb_service import ELBv2
+        from providers.aws.services.elbv2.elbv2_service import ELBv2
 
         current_audit_info.audited_partition = "aws"
 
         with mock.patch(
-            "providers.aws.services.elb.elbv2_request_smugling.elbv2_request_smugling.elbv2_client",
+            "providers.aws.services.elbv2.elbv2_desync_mitigation_mode.elbv2_desync_mitigation_mode.elbv2_client",
             new=ELBv2(current_audit_info),
         ):
-            from providers.aws.services.elb.elbv2_request_smugling.elbv2_request_smugling import (
-                elbv2_request_smugling,
+            from providers.aws.services.elbv2.elbv2_desync_mitigation_mode.elbv2_desync_mitigation_mode import (
+                elbv2_desync_mitigation_mode,
             )
 
-            check = elbv2_request_smugling()
+            check = elbv2_desync_mitigation_mode()
             result = check.execute()
 
             assert len(result) == 1
             assert result[0].status == "PASS"
             assert search(
-                "is dropping invalid header fields",
+                "is configured with correct desync mitigation mode",
                 result[0].status_extended,
             )
             assert result[0].resource_id == "my-lb"
