+++ providers/aws/services/elbv2/elbv2_listeners_underneath/elbv2_listeners_underneath.py
@@ -1,5 +1,5 @@
 from lib.check.models import Check, Check_Report
-from providers.aws.services.elb.elbv2_client import elbv2_client
+from providers.aws.services.elbv2.elbv2_client import elbv2_client
 
 
 class elbv2_listeners_underneath(Check):
