+++ providers/aws/services/elbv2/elbv2_listeners_underneath/elbv2_listeners_underneath_test.py
@@ -8,32 +8,32 @@ AWS_REGION = "eu-west-1"
 AWS_ACCOUNT_NUMBER = "123456789012"
 
 
-class Test_elbv2_desync_mitigation_mode:
+class Test_elbv2_listeners_underneath:
     @mock_elbv2
     def test_elb_no_balancers(self):
 
         from providers.aws.lib.audit_info.audit_info import current_audit_info
-        from providers.aws.services.elb.elb_service import ELBv2
+        from providers.aws.services.elbv2.elbv2_service import ELBv2
 
         current_audit_info.audited_partition = "aws"
 
         with mock.patch(
-            "providers.aws.services.elb.elbv2_desync_mitigation_mode.elbv2_desync_mitigation_mode.elbv2_client",
+            "providers.aws.services.elbv2.elbv2_listeners_underneath.elbv2_listeners_underneath.elbv2_client",
             new=ELBv2(current_audit_info),
         ):
             # Test Check
-            from providers.aws.services.elb.elbv2_desync_mitigation_mode.elbv2_desync_mitigation_mode import (
-                elbv2_desync_mitigation_mode,
+            from providers.aws.services.elbv2.elbv2_listeners_underneath.elbv2_listeners_underneath import (
+                elbv2_listeners_underneath,
             )
 
-            check = elbv2_desync_mitigation_mode()
+            check = elbv2_listeners_underneath()
             result = check.execute()
 
             assert len(result) == 0
 
     @mock_ec2
     @mock_elbv2
-    def test_elbv2_without_desync_mitigation_mode(self):
+    def test_elbv2_without_listeners(self):
         conn = client("elbv2", region_name=AWS_REGION)
         ec2 = resource("ec2", region_name=AWS_REGION)
 
@@ -56,33 +56,26 @@ class Test_elbv2_desync_mitigation_mode:
             Type="application",
         )["LoadBalancers"][0]
 
-        conn.modify_load_balancer_attributes(
-            LoadBalancerArn=lb["LoadBalancerArn"],
-            Attributes=[
-                {"Key": "routing.http.desync_mitigation_mode", "Value": "monitor"},
-            ],
-        )
-
         from providers.aws.lib.audit_info.audit_info import current_audit_info
-        from providers.aws.services.elb.elb_service import ELBv2
+        from providers.aws.services.elbv2.elbv2_service import ELBv2
 
         current_audit_info.audited_partition = "aws"
 
         with mock.patch(
-            "providers.aws.services.elb.elbv2_desync_mitigation_mode.elbv2_desync_mitigation_mode.elbv2_client",
+            "providers.aws.services.elbv2.elbv2_listeners_underneath.elbv2_listeners_underneath.elbv2_client",
             new=ELBv2(current_audit_info),
         ):
-            from providers.aws.services.elb.elbv2_desync_mitigation_mode.elbv2_desync_mitigation_mode import (
-                elbv2_desync_mitigation_mode,
+            from providers.aws.services.elbv2.elbv2_listeners_underneath.elbv2_listeners_underneath import (
+                elbv2_listeners_underneath,
             )
 
-            check = elbv2_desync_mitigation_mode()
+            check = elbv2_listeners_underneath()
             result = check.execute()
 
             assert len(result) == 1
             assert result[0].status == "FAIL"
             assert search(
-                "does not have desync mitigation mode set as defensive or strictest",
+                "has no listeners underneath",
                 result[0].status_extended,
             )
             assert result[0].resource_id == "my-lb"
@@ -90,7 +83,7 @@ class Test_elbv2_desync_mitigation_mode:
 
     @mock_ec2
     @mock_elbv2
-    def test_elbv2_with_desync_mitigation_mode(self):
+    def test_elbv2_with_listeners(self):
         conn = client("elbv2", region_name=AWS_REGION)
         ec2 = resource("ec2", region_name=AWS_REGION)
 
@@ -134,34 +127,24 @@ class Test_elbv2_desync_mitigation_mode:
             DefaultActions=[{"Type": "forward", "TargetGroupArn": target_group_arn}],
         )
 
-        conn.modify_load_balancer_attributes(
-            LoadBalancerArn=lb["LoadBalancerArn"],
-            Attributes=[
-                {"Key": "routing.http.desync_mitigation_mode", "Value": "defensive"},
-            ],
-        )
-
         from providers.aws.lib.audit_info.audit_info import current_audit_info
-        from providers.aws.services.elb.elb_service import ELBv2
+        from providers.aws.services.elbv2.elbv2_service import ELBv2
 
         current_audit_info.audited_partition = "aws"
 
         with mock.patch(
-            "providers.aws.services.elb.elbv2_desync_mitigation_mode.elbv2_desync_mitigation_mode.elbv2_client",
+            "providers.aws.services.elbv2.elbv2_listeners_underneath.elbv2_listeners_underneath.elbv2_client",
             new=ELBv2(current_audit_info),
         ):
-            from providers.aws.services.elb.elbv2_desync_mitigation_mode.elbv2_desync_mitigation_mode import (
-                elbv2_desync_mitigation_mode,
+            from providers.aws.services.elbv2.elbv2_listeners_underneath.elbv2_listeners_underneath import (
+                elbv2_listeners_underneath,
             )
 
-            check = elbv2_desync_mitigation_mode()
+            check = elbv2_listeners_underneath()
             result = check.execute()
 
             assert len(result) == 1
             assert result[0].status == "PASS"
-            assert search(
-                "is configured with correct desync mitigation mode",
-                result[0].status_extended,
-            )
+            assert search("has listeners underneath", result[0].status_extended)
             assert result[0].resource_id == "my-lb"
             assert result[0].resource_arn == lb["LoadBalancerArn"]
