+++ providers/aws/services/elbv2/elbv2_request_smugling/elbv2_request_smugling.py
@@ -1,5 +1,5 @@
 from lib.check.models import Check, Check_Report
-from providers.aws.services.elb.elbv2_client import elbv2_client
+from providers.aws.services.elbv2.elbv2_client import elbv2_client
 
 
 class elbv2_request_smugling(Check):
